+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+import useSafeButtonProps from './use-safe-button-props.hook';
+
+export { useSafeButtonProps };