+++ src/index.ts
@@ -1,11 +1,13 @@
 export * from '@/helpers';
 
+export * from '@/hooks';
+
 export * from '@/theme';
 
 export * from '@/ui/components';
 export * from '@/ui/forms';
 export * from '@/ui/provider';
 
-export type * from '@/types';
-
 export * from '@/utils';
+
+export type * from '@/types';