+++ src/types/ui/base/dom.types.ts
@@ -1,0 +1,2 @@
+export type ButtonEvent = React.MouseEvent<HTMLButtonElement, MouseEvent>;
+export type ButtonEventHandler = React.MouseEventHandler<HTMLButtonElement>;