+++ src/types/ui/base/index.ts
@@ -1,2 +1,3 @@
 export type * from './constants.types';
+export type * from './dom.types';
 export type * from './error.types';