+++ src/ui/forms/radio/radio.stories.tsx
@@ -45,6 +45,9 @@ const meta: Meta<typeof Radio> = {
       options: ['left', 'right'],
       control: { type: 'radio' },
       table: {
+        type: {
+          summary: 'string'
+        },
         defaultValue: { summary: 'left' }
       }
     },
@@ -68,6 +71,9 @@ const meta: Meta<typeof Radio> = {
       options: ['xs', 'sm', 'md', 'lg'],
       control: { type: 'select' },
       table: {
+        type: {
+          summary: 'string'
+        },
         defaultValue: { summary: 'md' }
       }
     }